move intentIQ analytics to its own TS module for clarity
