+++ bundle/src/lib/header-bidding/prebid/analytics.ts
@@ -1,13 +1,5 @@
-import type { ConsentState } from '@guardian/libs';
-import { getConsentFor } from '@guardian/libs';
 import type { AnalyticsConfig } from 'prebid.js/dist/libraries/analyticsAdapter/AnalyticsAdapter';
 import { getParticipations } from '../../../ab-testing';
-import {
-	EU_PARTNER_ID,
-	isUserInAllowedEURegion,
-	isUserInIntentIQRegion,
-	NON_EU_PARTNER_ID,
-} from './id-handlers/intent-iq';
 
 const shouldEnableAnalytics = (): boolean => {
 	if (!window.guardian.config.switches.prebidAnalytics) {
@@ -55,29 +47,6 @@ export const getGUAnalyticsConfig = (): AnalyticsConfig<'gu'> | undefined => {
 					window.location.hostname.includes('localhost')
 						? `//performance-events.code.dev-guardianapis.com/header-bidding`
 						: `//performance-events.guardianapis.com/header-bidding`,
-			},
-		};
-	}
-};
-
-export const getIntentIQAnalyticsConfig = (
-	consentState: ConsentState,
-): AnalyticsConfig<'iiqAnalytics'> | undefined => {
-	if (getConsentFor('intentIQ', consentState) && isUserInIntentIQRegion()) {
-		return {
-			provider: 'iiqAnalytics',
-			options: {
-				partner: isUserInAllowedEURegion()
-					? EU_PARTNER_ID
-					: NON_EU_PARTNER_ID,
-				ABTestingConfigurationSource: 'IIQServer',
-				domainName: 'theguardian.com',
-				gamObjectReference: googletag,
-				...(isUserInAllowedEURegion() && {
-					reportingServerAddress:
-						'https://reports-gdpr.intentiq.com/report',
-					browserBlackList: 'chrome',
-				}),
 			},
 		};
 	}
